Updated the JsrStepExecutionSplitter to provide a copy of the JobExecution per partition

+++ spring-batch-core/src/test/java/org/springframework/batch/core/scope/context/JobContextTests.java
@@ -180,7 +180,7 @@ public class JobContextTests {
 
 	@Test(expected = IllegalStateException.class)
 	public void testIllegalContextId() throws Exception {
-		context = new JobContext(new JobExecution(null));
+		context = new JobContext(new JobExecution((Long) null));
 		context.getId();
 	}
